Fixed bug

+++ scripts/generate_tcad_preinit_instructions.py
@@ -18,7 +18,7 @@ def main():
 
     log2_m = int(args.log2_m)
     training_length = int(args.training_length)
-    sensitivity_coefficient = int(args.sensitivity_coefficient)
+    sensitivity_coefficient = float(args.sensitivity_coefficient)
     input_log_file = args.input_log_file
    # output_rule_file = args.output_rule_file
